+++ packages/keyring-api/src/eth/utils.test.ts
@@ -4,7 +4,6 @@ import { isEvmAccountType } from './utils';
 describe('isEvmAccountType', () => {
   it.each([
     [EthAccountType.Eoa, true],
-    [EthAccountType.Erc4337, true],
     [BtcAccountType.P2pkh, false],
     [BtcAccountType.P2sh, false],
     [BtcAccountType.P2wpkh, false],